+++ src/components/Inputs/TextInput/MaskedTextInput.tsx
@@ -9,55 +9,55 @@ import type { MaskOptions } from 'react-native-mask-text/lib/typescript/@types/M
 
 
 
+// /** May be a function, where the param is the current value and it must return its mask. */
+export type Mask = string;// | ((p: {unmasked: string}) => string);
+// Removed function for now. Back then it was called outside this MaskedTextInput. Add later again inside it.
+// we used this on TextInputControlled: mask: typeof mask === 'function' ? mask({ unmasked }) : mask,
+
 type TIProps = Omit<RnTextInputProps, 'onChangeText'>;
 
 export type MaskedTextInputProps = TIProps & {
-  mask?: string;
-  type?: 'custom' | 'currency';
+  mask?: Mask;
+  maskType?: 'custom' | 'currency';
   options?: MaskOptions;
   defaultValue?: string;
   onChangeText?: (text: string, rawText: string) => void;
-  inputAccessoryView?: JSX.Element;
 };
 
 export const MaskedTextInput = forwardRef<RnTextInput, MaskedTextInputProps>(({
   mask: pattern,
-  type = 'custom',
+  maskType = 'custom',
   options = {} as MaskOptions,
   defaultValue: defaultValueProp,
   onChangeText,
-  inputAccessoryView,
   ...rest
 }, ref) => {
 
-  const defaultValue = defaultValueProp || (type === 'currency' ? '0' : '');
+  const defaultValue = defaultValueProp || (maskType === 'currency' ? '0' : '');
 
-  const [maskedValue, setMaskedValue] = useState(() => (pattern !== undefined ? mask(defaultValue, pattern, type, options) : defaultValue));
-  const [unMaskedValue, setUnmaskedValue] = useState(() => (pattern !== undefined ? unMask(defaultValue, type) : defaultValue));
+  const [maskedValue, setMaskedValue] = useState(() => (pattern !== undefined ? mask(defaultValue, pattern, maskType, options) : defaultValue));
+  const [unMaskedValue, setUnmaskedValue] = useState(() => (pattern !== undefined ? unMask(defaultValue, maskType) : defaultValue));
 
 
   const onChange = useCallback((value: string) => {
-    const newUnMaskedValue = pattern !== undefined ? unMask(value, type) : value;
-    const newMaskedValue = pattern !== undefined ? mask(newUnMaskedValue, pattern, type, options) : value;
+    const newUnMaskedValue = pattern !== undefined ? unMask(value, maskType) : value;
+    const newMaskedValue = pattern !== undefined ? mask(newUnMaskedValue, pattern, maskType, options) : value;
 
     setMaskedValue(newMaskedValue);
     setUnmaskedValue(newUnMaskedValue);
-  }, [options, pattern, type]);
+  }, [options, pattern, maskType]);
 
   useEffect(() => {
     onChangeText?.(maskedValue, unMaskedValue);
   }, [maskedValue, onChangeText, unMaskedValue]);
 
   return (
-    <>
-      <RnTextInput
-        onChangeText={(value) => onChange(value)}
-        ref={ref}
-        maxLength={pattern?.length || undefined}
-        {...rest}
-        value={maskedValue}
-      />
-      {inputAccessoryView}
-    </>
+    <RnTextInput
+      onChangeText={(value) => onChange(value)}
+      ref={ref}
+      maxLength={pattern?.length || undefined}
+      {...rest}
+      value={maskedValue}
+    />
   );
 });